requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-CORS==4.0.0
PyJWT==2.8.0
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.7
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
python-dateutil==2.8.2
pyotp==2.9.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
